observation space: change to continuous instead of discrete

+++ circle_environment.py
@@ -58,7 +58,7 @@ class CircleEnv(gym.Env):
         # We have 2 types of observations: the current state of the battery and the current distance to the next charging station
         # battery soc is in between 0 and 100000 Wh, distance to each of the charging stations is in between 0 and 2000 meters
         # -1 is used to signal that the vehicle is not spawned yet ("padding")
-        single_vehicle_observation_space = spaces.Box(low=np.array([-1, -1, -1, -1, -1]), high=np.array([100000, 2000, 2000, 2000, 2000]), dtype=int) 
+        single_vehicle_observation_space = spaces.Box(low=np.array([-1, -1, -1, -1, -1]), high=np.array([100000, 2000, 2000, 2000, 2000]), dtype=np.float32) 
         self.observation_space = spaces.Dict({
             str(vehicle_id): single_vehicle_observation_space for vehicle_id in self.vehicle_ids
         })
